+++ order/src/main/java/kitchenpos/order/application/TableService.java
@@ -2,7 +2,7 @@ package kitchenpos.order.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.order.dao.OrderTableRepository;
+import kitchenpos.order.repository.OrderTableRepository;
 import kitchenpos.order.domain.OrderTable;
 import kitchenpos.order.dto.OrderTableRequest;
 import kitchenpos.order.dto.OrderTableResponse;
