+++ product/src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.dao;
+package kitchenpos.product.repository;
 
 import kitchenpos.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
